+++ database/migrations/2015_02_17_152439_create_permission_user_table.php
@@ -15,9 +15,9 @@ class CreatePermissionUserTable extends Migration
     {
         Schema::create('permission_user', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('permission_id')->unsigned()->index();
+            $table->bigInteger('permission_id')->unsigned()->index();
             $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
         });
