+++ database/migrations/2015_02_28_164641_create_projects_table.php
@@ -17,7 +17,7 @@ class CreateProjectsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('stage')->nullable();
-            $table->integer('server_id')->unsigned();
+            $table->bigInteger('server_id')->unsigned();
             $table->string('repository');
             $table->string('email_notification_recipient')->nullable();
             $table->text('attributes');
@@ -25,7 +25,7 @@ class CreateProjectsTable extends Migration
             $table->integer('max_number_of_deployments_to_keep')->unsigned()->nullable();
             $table->tinyInteger('keep_last_deployment')->unsigned();
             $table->string('github_webhook_secret')->nullable();
-            $table->integer('github_webhook_user_id')->unsigned()->nullable();
+            $table->bigInteger('github_webhook_user_id')->unsigned()->nullable();
             $table->timestamps();
 
             $table->foreign('server_id')
